+++ qcengine/programs/util/__init__.py
@@ -1,2 +1,3 @@
 from .hessparse import load_hessian
 from .pdict import PreservingDict
+from .ao_reordering import reorder_column_ao_indices, reorder_row_and_column_ao_indices, cca_ao_order_spherical, get_ao_conversion
